requirements.txt
# LearnLive - TCP-based Classroom Management System
# Python Dependencies

# GUI Framework
ttkbootstrap==1.10.1

# Database
pymongo==4.6.0

# Email Notifications
secure-smtplib==0.1.1

# Utilities
python-dotenv==1.0.0
